move transformer embeddings to own file
